pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "node"
version = "0.1.0"
description = ""
authors = [ "moarshy <marshath@gmail.com>",]
readme = "README.md"
[[tool.poetry.packages]]
include = "node"
from = "."

[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "^23.2.1"
pydantic = "^2.5.3"
uvicorn = "^0.25.0"
fastapi = "0.100.0"
python-multipart = "^0.0.6"
python-dotenv = "^1.0.0"
httpx = "^0.26.0"
backoff = "^2.2.1"
docker = "^7.0.0"
celery = "^5.3.6"
redis = "^5.0.1"
surrealdb = "^0.3.2"
pyjwt = "^2.8.0"
psutil = "^5.9.7"
pytz = "^2023.3.post1"
gitpython = "^3.1.40"
poetry-core = "^1.8.1"
poetry = "^1.7.1"
click = "^8.1.7"
setuptools = "^69.0.3"
pyyaml = "^6.0.1"
protobuf = "^4.25.2"

[tool.poetry.scripts]
naptha = "cli:cli"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.11"
ipykernel = "^6.29.2"
